Add versions of the helper methods with explicit react
